--- a/src/QoiSharp/Codec/QoiCodec.cs
+++ b/src/QoiSharp/Codec/QoiCodec.cs
@@ -1,4 +1,6 @@
-﻿namespace QoiSharp.Codec;
+﻿using System.Buffers.Binary;
+
+namespace QoiSharp.Codec;
 
 /// <summary>
 /// QOI Codec.
@@ -32,8 +34,7 @@ public static class QoiCodec
     public static int CalculateHashTableIndex(int r, int g, int b, int a) =>
         ((r & 0xFF) * 3 + (g & 0xFF) * 5 + (b & 0xFF) * 7 + (a & 0xFF) * 11) % HashTableSize * 4;
 
-    public static bool IsValidMagic(byte[] magic) => CalculateMagic(magic) == Magic;
+    public static bool IsValidMagic(ReadOnlySpan<byte> magic) => BinaryPrimitives.ReadInt32BigEndian (magic) == Magic;
     
     private static int CalculateMagic(ReadOnlySpan<char> chars) => chars[0] << 24 | chars[1] << 16 | chars[2] << 8 | chars[3];
-    private static int CalculateMagic(ReadOnlySpan<byte> data) => data[0] << 24 | data[1] << 16 | data[2] << 8 | data[3];
 }