--- a/src/QoiSharp/QoiDecoder.cs
+++ b/src/QoiSharp/QoiDecoder.cs
@@ -129,14 +129,8 @@ public static class QoiDecoder
             }
         }
         
-        int pixelsEnd = data.Length - QoiCodec.Padding.Length;
-        for (int padIdx = 0; padIdx < QoiCodec.Padding.Length; padIdx++) 
-        {
-            if (data[pixelsEnd + padIdx] != QoiCodec.Padding[padIdx]) 
-            {
-                throw new InvalidOperationException("Invalid padding");
-            }
-        }
+        if (!QoiCodec.Padding.Span.SequenceEqual(data.Slice(data.Length - QoiCodec.Padding.Length)))
+            throw new InvalidOperationException("Invalid padding");
 
         return QoiImage.FromMemory(pixels, width, height, (Channels)channels, colorSpace);
     }