… tests/QoiSharp.Cli/Benchmarking/DecodingBenchmark.cs
… tests/QoiSharp.Cli/Benchmarking/EncodingBenchmark.cs
… tests/QoiSharp.Cli/Program.cs
… tests/QoiSharp.Tests/QoiTests.cs